requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.1
langchain-openai>=0.0.5
chromadb>=0.4.22
apscheduler>=3.10.4
pytz>=2023.3
pydantic>=2.5.0
aiohttp>=3.9.1
gunicorn>=21.2.0
crawl4ai>=0.2.0
playwright>=1.40.0